microservices-lab/auth-service/requirements.txt
psycopg2-binary==2.9.10
redis==4.5.5
python-dotenv==1.0.0
